refactor: splitted studyparticipant service into create, remove and lookup parts
