+++ src/Sepes.Infrastructure/Service/StudyParticipant/Interface/IStudyParticipantRemoveService.cs
@@ -1,0 +1,10 @@
+﻿using Sepes.Infrastructure.Dto.Study;
+using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Interface
+{
+    public interface IStudyParticipantRemoveService
+    {   
+        Task<StudyParticipantDto> RemoveAsync(int studyId, int userId, string roleName);
+    }
+}
